// block-performance-1.0.0/src/library/typescript/library-reference-list/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    library_name: string;
    recursive: boolean;
}

export interface OutputEmpty {
    library_list: string[];
    other_list: string[];
}

export interface OutputOk {
    library_list: string[];
    other_list: string[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: (output: OutputEmpty) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "library_name") return;
            if (field === "recursive") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.library_name must be string");
        if (typeof v.library_name !== "string") {
            throw new Error("v.library_name is not string");
        }

        log.println("v.recursive must be boolean");
        if (typeof v.recursive !== "boolean") {
            throw new Error("v.recursive is not boolean");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_empty(v: any): OutputEmpty {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_empty(v_cloned);
    } else {
        return _copy_output_empty(v);
    }
}

export function _copy_output_empty(v: any): OutputEmpty {
    if (typeof v === "object" && v !== null) {
        const obj = {
            library_list: copy_library_list(v.library_list),
            other_list: copy_other_list(v.other_list)
        };
        return obj;
    } else {
        return { library_list: [], other_list: [] };
    }

    function copy_library_list(v: any): string[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): string {
            return typeof v === "string" ? v : "";
        }
    }

    function copy_other_list(v: any): string[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): string {
            return typeof v === "string" ? v : "";
        }
    }
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            library_list: copy_library_list(v.library_list),
            other_list: copy_other_list(v.other_list)
        };
        return obj;
    } else {
        return { library_list: [], other_list: [] };
    }

    function copy_library_list(v: any): string[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): string {
            return typeof v === "string" ? v : "";
        }
    }

    function copy_other_list(v: any): string[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): string {
            return typeof v === "string" ? v : "";
        }
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;